A more reliable trigger for text size recalculation.

+++ src/EditorCell.h
@@ -478,7 +478,7 @@ private:
 
     We need to know this in order to be able to detect we need a full recalculation.
    */
-  int m_fontSize_Last;
+  double m_fontSize_Last;
   int m_charHeight;
   int m_paren1, m_paren2;
   //! Does this cell's size have to be recalculated?
